style(*): rename programdb model modules for PEP8
